+++ src/ballistica/scene_v1/support/client_session_net.h
@@ -26,8 +26,6 @@ class ClientSessionNet : public ClientSession {
 
  private:
   struct SampleBucket {
-    //    int least_buffered_count{};
-    //    int most_buffered_count{};
     int max_delay_from_projection{};
   };