+++ tools/bacommon/servermanager.py
@@ -143,6 +143,12 @@ class ServerConfig:
     # queue spamming attacks.
     enable_queue: bool = True
 
+    # Protocol version we host with. Currently the default is 33 which
+    # still allows older 1.4 game clients to connect. Explicitly setting
+    # to 35 no longer allows those clients but adds/fixes a few things
+    # such as making camera shake properly work in net games.
+    protocol_version: int | None = None
+
     # (internal) stress-testing mode.
     stress_test_players: int | None = None